requirements.txt
fastapi
uvicorn
requests
python-dotenv
apscheduler
sqlalchemy
aiohttp
pydantic
shodan
reportlab